2290 small refactoring
