+++ src/test/kotlin/com/epam/brn/controller/AdminExerciseFlowControllerTest.kt
@@ -1,12 +1,8 @@
 package com.epam.brn.controller
 
-import com.epam.brn.auth.AuthorityService
-import com.epam.brn.dto.response.BaseResponse
 import com.epam.brn.dto.ExerciseDto
-import com.epam.brn.dto.response.ExerciseWithTasksResponse
 import com.epam.brn.dto.ResourceDto
 import com.epam.brn.dto.StudyHistoryDto
-import com.epam.brn.dto.response.SubGroupResponse
 import com.epam.brn.dto.request.SubGroupChangeRequest
 import com.epam.brn.dto.request.SubGroupRequest
 import com.epam.brn.dto.request.UpdateResourceDescriptionRequest
@@ -15,20 +11,13 @@ import com.epam.brn.dto.request.exercise.ExerciseSentencesCreateDto
 import com.epam.brn.dto.request.exercise.ExerciseWordsCreateDto
 import com.epam.brn.dto.request.exercise.Phrases
 import com.epam.brn.dto.request.exercise.SetOfWords
-import com.epam.brn.dto.response.AuthorityResponse
-import com.epam.brn.dto.response.UserAccountResponse
-import com.epam.brn.dto.response.UserWithAnalyticsResponse
-import com.epam.brn.dto.statistic.DayStudyStatistic
-import com.epam.brn.dto.statistic.MonthStudyStatistic
+import com.epam.brn.dto.response.ExerciseWithTasksResponse
+import com.epam.brn.dto.response.SubGroupResponse
 import com.epam.brn.enums.Locale
-import com.epam.brn.enums.Role.ROLE_USER
 import com.epam.brn.service.ExerciseService
 import com.epam.brn.service.ResourceService
 import com.epam.brn.service.StudyHistoryService
-import com.epam.brn.service.UserAccountService
-import com.epam.brn.service.UserAnalyticsService
 import com.epam.brn.service.SubGroupService
-import com.epam.brn.service.statistic.UserPeriodStatisticService
 import com.epam.brn.upload.CsvUploadService
 import io.kotest.matchers.shouldBe
 import io.mockk.Runs
@@ -43,26 +32,13 @@ import io.mockk.verify
 import org.apache.http.HttpStatus
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
-import org.springframework.data.domain.Pageable
 import org.springframework.web.multipart.MultipartFile
 
 @ExtendWith(MockKExtension::class)
-internal class AdminControllerTest {
+internal class AdminExerciseFlowControllerTest {
 
     @InjectMockKs
-    private lateinit var adminController: AdminController
-
-    @MockK
-    private lateinit var userAccountService: UserAccountService
-
-    @MockK
-    private lateinit var userAnalyticsService: UserAnalyticsService
-
-    @MockK
-    private lateinit var userDayStatisticService: UserPeriodStatisticService<DayStudyStatistic>
-
-    @MockK
-    private lateinit var userMonthStatisticService: UserPeriodStatisticService<MonthStudyStatistic>
+    private lateinit var adminExerciseFlowController: AdminExerciseFlowController
 
     @MockK
     private lateinit var studyHistoryService: StudyHistoryService
@@ -80,22 +56,10 @@ internal class AdminControllerTest {
     private lateinit var subGroupService: SubGroupService
 
     @MockK
-    private lateinit var authorityService: AuthorityService
-
-    @MockK
-    private lateinit var pageable: Pageable
-
-    @MockK
     private lateinit var file: MultipartFile
 
     @MockK
     private lateinit var request: UpdateResourceDescriptionRequest
-
-    @MockK
-    private lateinit var userWithAnalyticsResponse: UserWithAnalyticsResponse
-
-    @MockK
-    private lateinit var userAccountResponse: UserAccountResponse
 
     @MockK
     private lateinit var studyHistoryDto: StudyHistoryDto
@@ -106,47 +70,6 @@ internal class AdminControllerTest {
     @MockK
     private lateinit var resourceDto: ResourceDto
 
-    @MockK
-    private lateinit var dayStudyStatistic: DayStudyStatistic
-
-    @MockK
-    private lateinit var monthStudyStatistic: MonthStudyStatistic
-
-    @MockK
-    private lateinit var authorityResponse: AuthorityResponse
-
-    @Test
-    fun `getUsers should return users with statistic when withAnalytics is true`() {
-        // GIVEN
-        val withAnalytics = true
-        val role = ROLE_USER.name
-        every { userAnalyticsService.getUsersWithAnalytics(pageable, role) } returns listOf(userWithAnalyticsResponse)
-
-        // WHEN
-        val users = adminController.getUsers(withAnalytics, role, pageable)
-
-        // THEN
-        verify(exactly = 1) { userAnalyticsService.getUsersWithAnalytics(pageable, role) }
-        users.statusCodeValue shouldBe HttpStatus.SC_OK
-        (users.body as BaseResponse).data shouldBe listOf(userWithAnalyticsResponse)
-    }
-
-    @Test
-    fun `getUsers should return users when withAnalytics is false`() {
-        // GIVEN
-        val withAnalytics = false
-        val role = ROLE_USER.name
-        every { userAccountService.getUsers(pageable, role) } returns listOf(userAccountResponse)
-
-        // WHEN
-        val users = adminController.getUsers(withAnalytics, role, pageable)
-
-        // THEN
-        verify(exactly = 1) { userAccountService.getUsers(pageable, role) }
-        users.statusCodeValue shouldBe HttpStatus.SC_OK
-        (users.body as BaseResponse).data shouldBe listOf(userAccountResponse)
-    }
-
     @Test
     fun `getMonthHistories should return month histories`() {
         // GIVEN
@@ -156,7 +79,7 @@ internal class AdminControllerTest {
         every { studyHistoryService.getMonthHistories(userId, month, year) } returns listOf(studyHistoryDto)
 
         // WHEN
-        val monthHistories = adminController.getMonthHistories(userId, month, year)
+        val monthHistories = adminExerciseFlowController.getMonthHistories(userId, month, year)
 
         // THEN
         verify(exactly = 1) { studyHistoryService.getMonthHistories(userId, month, year) }
@@ -171,7 +94,7 @@ internal class AdminControllerTest {
         every { csvUploadService.loadExercises(seriesId, file) } just Runs
 
         // WHEN
-        val loadExercises = adminController.loadExercises(seriesId, file)
+        val loadExercises = adminExerciseFlowController.loadExercises(seriesId, file)
 
         // THEN
         loadExercises.statusCodeValue shouldBe HttpStatus.SC_CREATED
@@ -184,7 +107,7 @@ internal class AdminControllerTest {
         every { exerciseService.findExercisesWithTasksBySubGroup(subGroupId) } returns listOf(exerciseWithTasksResponse)
 
         // WHEN
-        val exercises = adminController.getExercisesBySubGroup(subGroupId)
+        val exercises = adminExerciseFlowController.getExercisesBySubGroup(subGroupId)
 
         // THEN
         verify(exactly = 1) { exerciseService.findExercisesWithTasksBySubGroup(subGroupId) }
@@ -201,7 +124,7 @@ internal class AdminControllerTest {
         every { resourceService.updateDescription(id, description) } returns resourceDto
 
         // WHEN
-        val updated = adminController.updateResourceDescription(id, request)
+        val updated = adminExerciseFlowController.updateResourceDescription(id, request)
 
         // THEN
         verify(exactly = 1) { resourceService.updateDescription(id, description) }
@@ -219,23 +142,10 @@ internal class AdminControllerTest {
         every { subGroupService.addSubGroupToSeries(subGroupRequest, seriesId) } returns subGroupResponse
 
         // WHEN
-        val createdSubGroup = adminController.addSubGroupToSeries(seriesId, subGroupRequest)
+        val createdSubGroup = adminExerciseFlowController.addSubGroupToSeries(seriesId, subGroupRequest)
 
         // THEN
         createdSubGroup.statusCodeValue shouldBe HttpStatus.SC_CREATED
-    }
-
-    @Test
-    fun `getRoles should return http status 200`() {
-        // GIVEN
-
-        every { authorityService.findAll() } returns listOf(authorityResponse)
-
-        // WHEN
-        val authorities = adminController.getRoles()
-
-        // THEN
-        authorities.statusCodeValue shouldBe HttpStatus.SC_OK
     }
 
     @Test
@@ -253,7 +163,7 @@ internal class AdminControllerTest {
         every { exerciseService.createExercise(exerciseWordsCreateDto) } returns exerciseDto
 
         // WHEN
-        val createdExercise = adminController.createExercise(exerciseWordsCreateDto)
+        val createdExercise = adminExerciseFlowController.createExercise(exerciseWordsCreateDto)
 
         // THEN
         verify(exactly = 1) { exerciseService.createExercise(exerciseWordsCreateDto) }
@@ -275,7 +185,7 @@ internal class AdminControllerTest {
         every { exerciseService.createExercise(exercisePhrasesCreateDto) } returns exerciseDto
 
         // WHEN
-        val createdExercise = adminController.createExercise(exercisePhrasesCreateDto)
+        val createdExercise = adminExerciseFlowController.createExercise(exercisePhrasesCreateDto)
 
         // THEN
         verify(exactly = 1) { exerciseService.createExercise(exercisePhrasesCreateDto) }
@@ -297,7 +207,7 @@ internal class AdminControllerTest {
         every { exerciseService.createExercise(exerciseSentencesCreateDto) } returns exerciseDto
 
         // WHEN
-        val createdExercise = adminController.createExercise(exerciseSentencesCreateDto)
+        val createdExercise = adminExerciseFlowController.createExercise(exerciseSentencesCreateDto)
 
         // THEN
         verify(exactly = 1) { exerciseService.createExercise(exerciseSentencesCreateDto) }
@@ -313,7 +223,7 @@ internal class AdminControllerTest {
         every { subGroupService.updateSubGroupById(subGroupId, subGroupChangeRequest) } returns updatedSubGroup
 
         // WHEN
-        val actual = adminController.updateSubGroupById(subGroupId, subGroupChangeRequest)
+        val actual = adminExerciseFlowController.updateSubGroupById(subGroupId, subGroupChangeRequest)
 
         // THEN
         actual.statusCode.value() shouldBe HttpStatus.SC_OK
